\name{knnplus-package}
\alias{knnplus-package}
\alias{knnplus}
\docType{package}
\title{
\packageTitle{knnplus}
}
\description{
\packageDescription{knnplus}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{knnplus}
\packageIndices{knnplus}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{knnplus}

Maintainer: \packageMaintainer{knnplus}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
